+++ test/ml_dsa_test.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024-2025 The OpenSSL Project Authors. All Rights Reserved.
+ * Copyright 2024-2026 The OpenSSL Project Authors. All Rights Reserved.
  *
  * Licensed under the Apache License 2.0 (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
@@ -617,6 +617,94 @@ err:
     return ret;
 }
 
+/*
+ * A context string parameter of {NULL, 0} is an empty context string, the same
+ * as a parameter with a non-NULL pointer and a length of 0 or no parameter at
+ * all.  Check that signatures made with any one of those verify with any other,
+ * and that the ML-DSA-MU digest agrees.
+ */
+static int ml_dsa_empty_context_test(void)
+{
+    int ret = 0;
+    const char *alg = "ML-DSA-44";
+    EVP_PKEY *key = NULL;
+    EVP_PKEY_CTX *sctx = NULL, *vctx = NULL;
+    EVP_SIGNATURE *sig_alg = NULL;
+    EVP_MD *md = NULL;
+    EVP_MD_CTX *mctx = NULL;
+    uint8_t *sig = NULL, *pub = NULL;
+    size_t sig_len = 0, pub_len = 0;
+    uint8_t mu1[64], mu2[64];
+    OSSL_PARAM null_ctx[2], mu_params[3];
+
+    null_ctx[0] = OSSL_PARAM_construct_octet_string(OSSL_SIGNATURE_PARAM_CONTEXT_STRING,
+        NULL, 0);
+    null_ctx[1] = OSSL_PARAM_construct_end();
+
+    if (!TEST_ptr(key = do_gen_key(alg, NULL, 0))
+        || !TEST_ptr(sig_alg = EVP_SIGNATURE_fetch(lib_ctx, alg, NULL)))
+        goto err;
+
+    /* Sign with {NULL, 0}, verify with the parameter omitted */
+    if (!TEST_ptr(sctx = EVP_PKEY_CTX_new_from_pkey(lib_ctx, key, NULL))
+        || !TEST_int_eq(EVP_PKEY_sign_message_init(sctx, sig_alg, null_ctx), 1)
+        || !TEST_int_eq(EVP_PKEY_sign(sctx, NULL, &sig_len, msg1, sizeof(msg1)), 1)
+        || !TEST_ptr(sig = OPENSSL_malloc(sig_len))
+        || !TEST_int_eq(EVP_PKEY_sign(sctx, sig, &sig_len, msg1, sizeof(msg1)), 1)
+        || !TEST_ptr(vctx = EVP_PKEY_CTX_new_from_pkey(lib_ctx, key, NULL))
+        || !TEST_int_eq(EVP_PKEY_verify_message_init(vctx, sig_alg, NULL), 1)
+        || !TEST_int_eq(EVP_PKEY_verify(vctx, sig, sig_len, msg1, sizeof(msg1)), 1))
+        goto err;
+    EVP_PKEY_CTX_free(sctx);
+    EVP_PKEY_CTX_free(vctx);
+    sctx = vctx = NULL;
+
+    /* Sign with the parameter omitted, verify with {NULL, 0} */
+    if (!TEST_ptr(sctx = EVP_PKEY_CTX_new_from_pkey(lib_ctx, key, NULL))
+        || !TEST_int_eq(EVP_PKEY_sign_message_init(sctx, sig_alg, NULL), 1)
+        || !TEST_int_eq(EVP_PKEY_sign(sctx, sig, &sig_len, msg1, sizeof(msg1)), 1)
+        || !TEST_ptr(vctx = EVP_PKEY_CTX_new_from_pkey(lib_ctx, key, NULL))
+        || !TEST_int_eq(EVP_PKEY_verify_message_init(vctx, sig_alg, null_ctx), 1)
+        || !TEST_int_eq(EVP_PKEY_verify(vctx, sig, sig_len, msg1, sizeof(msg1)), 1))
+        goto err;
+
+    /* ML-DSA-MU with {NULL, 0} gives the same mu as with no context string */
+    if (!TEST_int_eq(EVP_PKEY_get_raw_public_key(key, NULL, &pub_len), 1)
+        || !TEST_ptr(pub = OPENSSL_malloc(pub_len))
+        || !TEST_int_eq(EVP_PKEY_get_raw_public_key(key, pub, &pub_len), 1)
+        || !TEST_ptr(md = EVP_MD_fetch(lib_ctx, "ML-DSA-MU", NULL))
+        || !TEST_ptr(mctx = EVP_MD_CTX_new()))
+        goto err;
+    mu_params[0] = OSSL_PARAM_construct_octet_string(OSSL_DIGEST_PARAM_MU_PUB_KEY,
+        pub, pub_len);
+    mu_params[1] = OSSL_PARAM_construct_end();
+    if (!TEST_true(EVP_DigestInit_ex2(mctx, md, mu_params))
+        || !TEST_true(EVP_DigestUpdate(mctx, msg1, sizeof(msg1)))
+        || !TEST_true(EVP_DigestFinalXOF(mctx, mu1, sizeof(mu1))))
+        goto err;
+    mu_params[1] = OSSL_PARAM_construct_octet_string(OSSL_DIGEST_PARAM_MU_CONTEXT_STRING,
+        NULL, 0);
+    mu_params[2] = OSSL_PARAM_construct_end();
+    EVP_MD_CTX_free(mctx);
+    if (!TEST_ptr(mctx = EVP_MD_CTX_new())
+        || !TEST_true(EVP_DigestInit_ex2(mctx, md, mu_params))
+        || !TEST_true(EVP_DigestUpdate(mctx, msg1, sizeof(msg1)))
+        || !TEST_true(EVP_DigestFinalXOF(mctx, mu2, sizeof(mu2)))
+        || !TEST_mem_eq(mu1, sizeof(mu1), mu2, sizeof(mu2)))
+        goto err;
+    ret = 1;
+err:
+    EVP_MD_CTX_free(mctx);
+    EVP_MD_free(md);
+    EVP_PKEY_CTX_free(sctx);
+    EVP_PKEY_CTX_free(vctx);
+    EVP_SIGNATURE_free(sig_alg);
+    EVP_PKEY_free(key);
+    OPENSSL_free(sig);
+    OPENSSL_free(pub);
+    return ret;
+}
+
 static int ml_dsa_priv_pub_bad_t0_test(void)
 {
     int ret = 0;
@@ -796,6 +884,8 @@ int setup_tests(void)
     ADD_TEST(from_data_invalid_public_test);
     ADD_TEST(from_data_bad_input_test);
     ADD_TEST(ml_dsa_digest_sign_verify_test);
+    if (fips_provider_version_ge(lib_ctx, 4, 1, 0))
+        ADD_TEST(ml_dsa_empty_context_test);
     ADD_TEST(ml_dsa_priv_pub_bad_t0_test);
 
     /*