+++ test/slh_dsa_test.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024-2025 The OpenSSL Project Authors. All Rights Reserved.
+ * Copyright 2024-2026 The OpenSSL Project Authors. All Rights Reserved.
  *
  * Licensed under the Apache License 2.0 (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
@@ -585,6 +585,63 @@ err:
     return ret;
 }
 
+/*
+ * A context string parameter of {NULL, 0} is an empty context string, the same
+ * as no context string parameter at all.  Check that a signature made with one
+ * verifies with the other.
+ */
+static int slh_dsa_empty_context_test(void)
+{
+    int ret = 0;
+    const char *alg = "SLH-DSA-SHA2-128f";
+    EVP_PKEY *key = NULL;
+    EVP_PKEY_CTX *sctx = NULL, *vctx = NULL;
+    EVP_SIGNATURE *sig_alg = NULL;
+    uint8_t *sig = NULL;
+    size_t sig_len = 0;
+    static uint8_t msg[] = "Hello World";
+    OSSL_PARAM null_ctx[2];
+
+    null_ctx[0] = OSSL_PARAM_construct_octet_string(OSSL_SIGNATURE_PARAM_CONTEXT_STRING,
+        NULL, 0);
+    null_ctx[1] = OSSL_PARAM_construct_end();
+
+    if (!TEST_ptr(key = do_gen_key(alg, NULL, 0))
+        || !TEST_ptr(sig_alg = EVP_SIGNATURE_fetch(lib_ctx, alg, NULL)))
+        goto err;
+
+    /* Sign with {NULL, 0}, verify with the parameter omitted */
+    if (!TEST_ptr(sctx = EVP_PKEY_CTX_new_from_pkey(lib_ctx, key, NULL))
+        || !TEST_int_eq(EVP_PKEY_sign_message_init(sctx, sig_alg, null_ctx), 1)
+        || !TEST_int_eq(EVP_PKEY_sign(sctx, NULL, &sig_len, msg, sizeof(msg)), 1)
+        || !TEST_ptr(sig = OPENSSL_malloc(sig_len))
+        || !TEST_int_eq(EVP_PKEY_sign(sctx, sig, &sig_len, msg, sizeof(msg)), 1)
+        || !TEST_ptr(vctx = EVP_PKEY_CTX_new_from_pkey(lib_ctx, key, NULL))
+        || !TEST_int_eq(EVP_PKEY_verify_message_init(vctx, sig_alg, NULL), 1)
+        || !TEST_int_eq(EVP_PKEY_verify(vctx, sig, sig_len, msg, sizeof(msg)), 1))
+        goto err;
+    EVP_PKEY_CTX_free(sctx);
+    EVP_PKEY_CTX_free(vctx);
+    sctx = vctx = NULL;
+
+    /* Sign with the parameter omitted, verify with {NULL, 0} */
+    if (!TEST_ptr(sctx = EVP_PKEY_CTX_new_from_pkey(lib_ctx, key, NULL))
+        || !TEST_int_eq(EVP_PKEY_sign_message_init(sctx, sig_alg, NULL), 1)
+        || !TEST_int_eq(EVP_PKEY_sign(sctx, sig, &sig_len, msg, sizeof(msg)), 1)
+        || !TEST_ptr(vctx = EVP_PKEY_CTX_new_from_pkey(lib_ctx, key, NULL))
+        || !TEST_int_eq(EVP_PKEY_verify_message_init(vctx, sig_alg, null_ctx), 1)
+        || !TEST_int_eq(EVP_PKEY_verify(vctx, sig, sig_len, msg, sizeof(msg)), 1))
+        goto err;
+    ret = 1;
+err:
+    EVP_PKEY_CTX_free(sctx);
+    EVP_PKEY_CTX_free(vctx);
+    EVP_SIGNATURE_free(sig_alg);
+    EVP_PKEY_free(key);
+    OPENSSL_free(sig);
+    return ret;
+}
+
 static int slh_dsa_keygen_invalid_test(void)
 {
     int ret = 0;
@@ -670,6 +727,8 @@ int setup_tests(void)
     ADD_ALL_TESTS(slh_dsa_sign_verify_test, OSSL_NELEM(slh_dsa_sig_testdata));
     ADD_ALL_TESTS(slh_dsa_keygen_test, OSSL_NELEM(slh_dsa_keygen_testdata));
     ADD_TEST(slh_dsa_digest_sign_verify_test);
+    if (fips_provider_version_ge(lib_ctx, 4, 1, 0))
+        ADD_TEST(slh_dsa_empty_context_test);
     ADD_TEST(slh_dsa_keygen_invalid_test);
     return 1;
 }